requirements.txt
Django==2.2.4
django-jet==1.0.8
djangorestframework==3.10.3
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
